requirements.txt
openai
langchain
langchain-core
langchain-community
langchain-groq
pandas
matplotlib
seaborn
python-dotenv
dateparser
